+++ tests/TestCase/FieldHandlers/FieldHandlerFactoryTest.php
@@ -9,7 +9,8 @@ use Cake\TestSuite\TestCase;
 use CsvMigrations\FieldHandlers\CsvField;
 use CsvMigrations\FieldHandlers\DbField;
 use CsvMigrations\FieldHandlers\FieldHandlerFactory;
-use CsvMigrations\MigrationTrait;
+use CsvMigrations\Parser\Csv\MigrationParser;
+use CsvMigrations\PathFinder\MigrationPathFinder;
 
 /**
  * Foo Entity.
@@ -79,10 +80,14 @@ class FieldHandlerFactoryTest extends TestCase
     {
         parent::setUp();
 
-        $dir = dirname(__DIR__) . DS . '..' . DS . 'data' . DS . 'CsvMigrations' . DS;
+        $dir = dirname(__DIR__) . DS . '..' . DS . 'data' . DS . 'Modules' . DS;
+        Configure::write('CsvMigrations.modules.path', $dir);
 
-        $mockTrait = $this->getMockForTrait(MigrationTrait::class);
-        $this->csvData = $mockTrait->getFieldsDefinitions($this->tableName);
+        $pf = new MigrationPathFinder();
+        $path = $pf->find('Foo');
+        $parser = new MigrationParser();
+        $this->csvData = $parser->wrapFromPath($path);
+
         $config = TableRegistry::exists($this->tableName)
             ? []
             : ['className' => 'CsvMigrations\Test\TestCase\Model\Table\FooTable'];
